pyproject.toml
[tool.poetry]
name = "carpooling"
version = "0.1.0"
description = ""
authors = [
"Yousab Azmy <yousab.azmy@orange.com>",
"Mohamed Ismail <mohamed.g.ismail@orange.com>"
]
[tool.poetry.dependencies]
python = "~3.11"
Django = "~4.1"
pyyaml = "~6.0"
requests = "~2.31"
django-filter = "~23.2"
django-crispy-forms = "~2.0"
crispy-bootstrap4 = "^2022.1"
django-phonenumber-field = "^7.1.0"
django-phonenumbers = "^1.0.1"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
markdown = "^3.4.4"
drf-yasg = {extras = ["validation"], version = "^1.21.7"}
boto3 = "^1.28.27"
pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
bandit = "~1.7"
black = "~23.1"
pylint = "~2.17"
pytest = "~7.2"
pytest-cov = "~4.0"
pytest-django = "~4.5"
pytest-mock = "~3.10"
pytest-xdist = "~3.2"
safety = "~2.3"
pre-commit = "^3.3.2"
autopep8 = "^2.0.2"

[tool.poetry.group.prod]
optional = true

[tool.poetry.group.prod.dependencies]
daphne = "~4.0"
twisted = {version = "~22.10", extras = ["tls", "http2"]}
mysqlclient = "~2.1"

[[tool.poetry.source]]
name = "diod_python_proxy"
url = "https://repos.tech.orange/artifactory/api/pypi/pythonproxy/simple"
default = true
secondary = false

[tool.autoflake]
check = true
imports = ["django", "requests", "urllib3"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "bootstrap.settings.testing"
python_files = ["tests/**/test*.py"]
addopts = "-v"


[tool.bandit]
exclude_dirs = [".venv", ".cache", ".pytest_cache", "reports", "static", "tests", "nginx"]
tests = ["B201", "B301"]

[tool.flake8]
ignore = "E203, E266, E501, W503, F403, F401"
max-line-length = 79
max-complexity = 18
select = "B,C,E,F,W,T4,B9"


[tool.black]
line-length = 110
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | \.vscode
  | \.idea
)/
'''

[tool.isort]
multi_line_output = 3
line_length = 88
include_trailing_comma = true
